opa-gov-api/pyproject.toml
[project]
name = "opa-gov-api"
version = "0.1.0"
description = "FastAPI wrapper around OPA for inline security and masking decisions"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "python-dotenv>=1.0.1",
    "fastapi>=0.116.1",
    "httpx>=0.27.0",
    "uvicorn>=0.35.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-prometheus>=0.48b0",
    "prometheus-client>=0.20.0",
]
